+++ src/app/components/invoice/create-invoice/create-invoice.component.ts
@@ -1,13 +1,10 @@
 import { Component } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { AgGridModule } from 'ag-grid-angular';
-import { CellClassParams, GetRowIdParams, GridOptions, GridReadyEvent, ICellRendererParams, ValueFormatterParams } from 'ag-grid-community';
+import { GetRowIdParams, GridOptions } from 'ag-grid-community';
 import { FormColumnDef } from '../../../../util/form-column-def.type';
-import { ItemCellEditorComponent } from '../../common/item-cell-editor/item-cell-editor.component';
 import { loadCountries } from '../store/actions/country.actions';
 import { CreateInvoiceCustomerComponent } from './create-invoice-customer.component';
-
-
 @Component({
   selector: 'app-create-invoice',
   standalone: true,
@@ -37,70 +34,11 @@ export class CreateInvoiceComponent extends CreateInvoiceCustomerComponent {
     editable: false
   };
 
-  itemdetailsColumnDefs = [
-    {
-      headerName: 'Item',
-      field: 'item',
-      flex: 2,
-      editable: true,
-      cellEditor: ItemCellEditorComponent,
-      valueFormatter: (params: ValueFormatterParams) => {
-        return params.value && params.value.trim() !== ''
-          ? params.value
-          : 'Enter item here';
-      },
+  
 
-      cellClass: (params: CellClassParams) => {
-        return !params.value || params.value.trim() === '' ? 'item-placeholder' : '';
-      }
-
-    },
-    { headerName: 'Price', field: 'price', flex: 1, editable: true, cellClass: 'right-align' },
-    { headerName: 'Quantity', field: 'quantity', flex: 1, editable: true, cellClass: 'right-align' },
-    { headerName: 'Total', field: 'total', flex: 1, cellClass: 'right-align' },
-    {
-      headerName: '',
-      field: 'delete',
-      flex: 0.5,
-      cellRenderer: (params: ICellRendererParams) => {
-        const rowCount = params.api.getDisplayedRowCount();
-        const isLastRow = rowCount === 1;
-
-        if (isLastRow) return '';
-
-        return `
-    <button class="delete-btn" data-clear="true" aria-label="Delete">
-      <span class="material-icons" style="color: red;">delete</span>
-    </button>
-  `;
-      },
-      suppressMenu: true,
-      suppressSorting: true
-    }
-  ];
-  itemdetailsRowData = [
-    { item: '', price: '', quantity: '', total: '' }
-  ];
-
-  summaryColumnDefs: ColDef<FormColumnDef>[] = [
-    { field: 'label', headerName: '', width: 150 },
-    { field: 'value', headerName: '', width: 200, editable: true }
-  ];
-
-  summaryRowData: FormColumnDef[] = [
-    { label: 'Item Total', value: '' },
-    { label: 'Discount', value: '' },
-    { label: 'Sub Total', value: '' },
-    { label: 'Tax', value: '' },
-    { label: 'Rount Off', value: '' },
-    { label: 'Grand Total', value: '' }
-  ];
+  
 
 
-  onCellClicked(event: any): void {
-    const isDeleteBtn = event.event?.target?.closest('[data-clear]');
-    if (event.colDef.field === 'delete' && isDeleteBtn) {
-      const index = event.rowIndex;
   onCellClicked(event: any): void {
     const isDeleteBtn = event.event?.target?.closest('[data-clear]');
     if (event.colDef.field === 'delete' && isDeleteBtn) {
@@ -114,11 +52,6 @@ export class CreateInvoiceComponent extends CreateInvoiceCustomerComponent {
   }
 
 
-  onCellValueChanged(event: any): void {
-    const data = event.data;
-    const rowIndex = event.rowIndex;
-    const price = Number(data.price);
-    let quantity = Number(data.quantity);
   onCellValueChanged(event: any): void {
     const data = event.data;
     const rowIndex = event.rowIndex;
@@ -150,23 +83,13 @@ export class CreateInvoiceComponent extends CreateInvoiceCustomerComponent {
   onCellKeyDown(event: any): void {
     const isEnter = event.event.key === 'Enter';
     if (!isEnter) return;
-  onCellKeyDown(event: any): void {
-    const isEnter = event.event.key === 'Enter';
-    if (!isEnter) return;
 
     const row = event.node.data;
-    const row = event.node.data;
-
-    const isComplete =
-      row.item?.trim() &&
-      Number(row.price) > 0 &&
-      Number(row.quantity) > 0;
     const isComplete =
       row.item?.trim() &&
       Number(row.price) > 0 &&
       Number(row.quantity) > 0;
 
-    const isLastRow = event.rowIndex === this.itemdetailsRowData.length - 1;
     const isLastRow = event.rowIndex === this.itemdetailsRowData.length - 1;
 
     if (isComplete && isLastRow) {
